% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_operations.R
\name{copy_db_parameter_group}
\alias{copy_db_parameter_group}
\title{Copies the specified DB parameter group}
\usage{
copy_db_parameter_group(SourceDBParameterGroupIdentifier,
  TargetDBParameterGroupIdentifier, TargetDBParameterGroupDescription,
  Tags = NULL)
}
\arguments{
\item{SourceDBParameterGroupIdentifier}{The identifier or ARN for the source DB parameter group. For information about creating an ARN, see \href{http://docs.aws.amazon.com/neptune/latest/UserGuide/tagging.ARN.html#tagging.ARN.Constructing}{Constructing an Amazon Resource Name (ARN)}.

Constraints:
\itemize{
\item Must specify a valid DB parameter group.
\item Must specify a valid DB parameter group identifier, for example \code{my-db-param-group}, or a valid ARN.
}}

\item{TargetDBParameterGroupIdentifier}{The identifier for the copied DB parameter group.

Constraints:
\itemize{
\item Cannot be null, empty, or blank
\item Must contain from 1 to 255 letters, numbers, or hyphens
\item First character must be a letter
\item Cannot end with a hyphen or contain two consecutive hyphens
}

Example: \code{my-db-parameter-group}}

\item{TargetDBParameterGroupDescription}{A description for the copied DB parameter group.}

\item{Tags}{}
}
\description{
Copies the specified DB parameter group.
}
\examples{

}
