% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_operations.R
\name{delete_event_subscription}
\alias{delete_event_subscription}
\title{Deletes an event notification subscription}
\usage{
delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the event notification subscription you want to delete.}
}
\description{
Deletes an event notification subscription.
}
\section{Accepted Parameters}{
\preformatted{delete_event_subscription(
  SubscriptionName = "string"
)
}
}

