% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_operations.R
\name{describe_event_categories}
\alias{describe_event_categories}
\title{Displays a list of categories for all event source types, or, if specified, for a specified source type}
\usage{
describe_event_categories(SourceType = NULL, Filters = NULL)
}
\arguments{
\item{SourceType}{The type of source that is generating the events.

Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot}

\item{Filters}{This parameter is not currently supported.}
}
\description{
Displays a list of categories for all event source types, or, if specified, for a specified source type.
}
\examples{

}
