% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists all tags on an Amazon Neptune resource}
\usage{
list_tags_for_resource(
  ResourceName = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}
\arguments{
\item{ResourceName}{[required] The Amazon Neptune resource with tags to be listed. This value is an Amazon Resource Name (ARN). For information about creating an ARN, see \href{http://docs.aws.amazon.com/neptune/latest/UserGuide/tagging.ARN.html#tagging.ARN.Constructing}{Constructing an Amazon Resource Name (ARN)}.}

\item{Filters}{This parameter is not currently supported.}
}
\description{
Lists all tags on an Amazon Neptune resource.
}
\examples{

}
