% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_package.R
\docType{package}
\name{paws.neptune-package}
\alias{paws.neptune}
\alias{paws.neptune-package}
\title{paws.neptune: Amazon Neptune}
\description{
Amazon Neptune is a fast, reliable, fully-managed
    graph database service that makes it easy to build and run
    applications that work with highly connected datasets. The core of
    Amazon Neptune is a purpose-built, high-performance graph database
    engine optimized for storing billions of relationships and querying
    the graph with milliseconds latency. Amazon Neptune supports popular
    graph models Property Graph and W3C\'s RDF, and their respective query
    languages Apache TinkerPop Gremlin and SPARQL, allowing you to easily
    build queries that efficiently navigate highly connected datasets.
    Neptune powers graph use cases such as recommendation engines, fraud
    detection, knowledge graphs, drug discovery, and network security.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
