% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_operations.R
\name{reboot_db_instance}
\alias{reboot_db_instance}
\title{You might need to reboot your DB instance, usually for maintenance reasons}
\usage{
reboot_db_instance(DBInstanceIdentifier, ForceFailover = NULL)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:
\itemize{
\item Must match the identifier of an existing DBInstance.
}}

\item{ForceFailover}{When \code{true}, the reboot is conducted through a MultiAZ failover.

Constraint: You can't specify \code{true} if the instance is not configured for MultiAZ.}
}
\description{
You might need to reboot your DB instance, usually for maintenance reasons. For example, if you make certain modifications, or if you change the DB parameter group associated with the DB instance, you must reboot the instance for the changes to take effect.
}
\details{
Rebooting a DB instance restarts the database engine service. Rebooting a DB instance results in a momentary outage, during which the DB instance status is set to rebooting.
}
\section{Accepted Parameters}{
\preformatted{reboot_db_instance(
  DBInstanceIdentifier = "string",
  ForceFailover = TRUE|FALSE
)
}
}

