% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_operations.R
\name{remove_role_from_db_cluster}
\alias{remove_role_from_db_cluster}
\title{Disassociates an Identity and Access Management (IAM) role from a DB cluster}
\usage{
remove_role_from_db_cluster(DBClusterIdentifier, RoleArn)
}
\arguments{
\item{DBClusterIdentifier}{The name of the DB cluster to disassociate the IAM role from.}

\item{RoleArn}{The Amazon Resource Name (ARN) of the IAM role to disassociate from the DB cluster, for example \code{arn:aws:iam::123456789012:role/NeptuneAccessRole}.}
}
\description{
Disassociates an Identity and Access Management (IAM) role from a DB cluster.
}
\examples{

}
