% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_operations.R
\name{remove_source_identifier_from_subscription}
\alias{remove_source_identifier_from_subscription}
\title{Removes a source identifier from an existing event notification subscription}
\usage{
remove_source_identifier_from_subscription(SubscriptionName, SourceIdentifier)
}
\arguments{
\item{SubscriptionName}{The name of the event notification subscription you want to remove a source identifier from.}

\item{SourceIdentifier}{The source identifier to be removed from the subscription, such as the \strong{DB instance identifier} for a DB instance or the name of a security group.}
}
\description{
Removes a source identifier from an existing event notification subscription.
}
\examples{

}
