% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.neptune_operations.R
\name{remove_tags_from_resource}
\alias{remove_tags_from_resource}
\title{Removes metadata tags from an Amazon Neptune resource}
\usage{
remove_tags_from_resource(ResourceName, TagKeys)
}
\arguments{
\item{ResourceName}{The Amazon Neptune resource that the tags are removed from. This value is an Amazon Resource Name (ARN). For information about creating an ARN, see \href{http://docs.aws.amazon.com/neptune/latest/UserGuide/tagging.ARN.html#tagging.ARN.Constructing}{Constructing an Amazon Resource Name (ARN)}.}

\item{TagKeys}{The tag key (name) of the tag to be removed.}
}
\description{
Removes metadata tags from an Amazon Neptune resource.
}
\examples{

}
